refactor(frontend): decompose nodeCards.tsx into per-kind files (PR5h.12 review)
Final PR5h reviewer must-fix. nodeCards.tsx had grown to 752 LOC
with 6 cards + CardFrame + 4 inline helpers + 3 inline editors + 4
FanCard sub-components — review surface for any future PR6/PR7 touch
was prohibitive. Pure mechanical extraction; no behavior changes; all
tests pass unchanged.

## What ships

Eight files, one barrel:

1. `cardFrame.tsx` — `CardFrame`, `CardBody`, `MetaRow`, `CardFrameProps`
   (the shared building blocks every card uses).
2. `RootPromptCard.tsx` — `RootPromptCard` + private `InlineRootPromptEditor`.
3. `ImportMessageCard.tsx` — `ImportMessageCard`.
4. `UserTurnCard.tsx` — `UserTurnCard` + exported `ConverterChipRow` +
   exported `InlineTextEditor` + private `converterLabel`.
5. `SendCard.tsx` — `SendCard` (with the inline `lastError` panel).
6. `FanCard.tsx` — `FanCard` + private `StackSummaryBody` +
   `StackToggleButton` + `StackPickButton`.
7. `ScoreCard.tsx` — `ScoreCard`.
8. `nodeCards.tsx` (REWRITTEN as 16-line barrel) — re-exports the 6
   card components. Existing imports (`treeNodeTypes.ts`, every test
   file) stay stable.

## Notable shape decisions

- `CardFrame` is **not** re-exported through nodeCards barrel (was
  never publicly exposed). It's exported from `cardFrame.tsx` for
  per-kind card files to import.
- `ConverterChipRow` and `InlineTextEditor` are exported from
  `UserTurnCard.tsx` — both are components and could plausibly be
  reused by other cards in V1.0.1 (e.g., SendCard might want to
  display its retry attempts as chips). Exporting them now is
  cheap and forward-looking.
- `converterLabel` is **private** (`function`, not `export function`)
  per `react-refresh/only-export-components` lint rule: a file can
  only export components, not bare helper functions. Moving it out
  would require a separate file for one tiny helper; keeping it
  module-private is the right scope for V1.0.
- `InlineRootPromptEditor` stays private to `RootPromptCard.tsx`
  (no other card uses it; if RootPrompt's edit shape changes, the
  editor is updated alongside).
- Each per-kind file imports `CardFrame`/`CardBody`/`MetaRow` from
  `./cardFrame`, not from a shared barrel. Avoids circular imports
  if a card ever depends on another card.
- All in-file comments preserved verbatim — the helper JSDocs that
  describe StackSummaryBody, StackPickButton, ConverterChipRow,
  InlineTextEditor, InlineRootPromptEditor stay with their parent
  files.

## TDD narrative

- This is a pure mechanical refactor; no new tests, no test edits.
- Sequence: created 7 new files via the test-then-implement pattern
  but inverted (read the original file as the source of truth, split
  by section markers in nodeCards.tsx, paste verbatim into per-kind
  files with adjusted imports). Used a subagent for the parallel
  split — the subagent produced the 8 file contents; the parent
  created the files + ran verification.
- Verified by running `npm test --no-coverage` after each file
  creation step; final run: 1112 passed (was 1112 before
  decomposition — identical).

## Defects surfaced during TDD

- The subagent's initial subdivision exported `converterLabel`
  (per my prompt's "export them" guidance for ConverterChipRow +
  converterLabel + InlineTextEditor). ESLint flagged
  `react-refresh/only-export-components` — bare functions cannot be
  exported from a component file. Demoted `converterLabel` to a
  private module function. The other two exports are components and
  remain.
- One leftover terminal eslint-disable directive in nodeCards.tsx
  (line 752 of the old file) became moot when the file shrank to 16
  lines.

## Verification

- `npm test --no-coverage`: 1112 passed (unchanged; no test edits)
- `npm run type-check`: 0
- `npm run type-check:contract`: 0
- `npm run lint`: 0 warnings
- nodeCards.tsx line count: 752 → 16

## Next slice

PR5h sequence is COMPLETE. All 5 PR5h reviewer must-fixes landed:
- ✅ #1 nodeCards.tsx decomposition (THIS COMMIT)
- ✅ #2 toFlowEdge silent default (PR5h.8)
- ✅ #3 InlineRootPromptEditor Cmd-Enter inconsistency (PR5h.9)
- ✅ #4 converter remove chip (PR5h.10)
- ✅ #5 actionCallbacks memoization (PR5h.11)

Plus all PR5h non-blocking items:
- ✅ findCard test helper extracted (PR5h.8)
- ✅ void screen/userEvent dropped (PR5h.8)
- ✅ tooltip test renamed (PR5h.8)
- ✅ WHY comment on useShapeMemoizedLayout (PR5h.8)

Next: PR6 (wave UX). Sub-PRs:
- PR6a: cost-guardrail modal
- PR6b: ↻ cost-preview tooltip
- PR6c: wave-status ribbon + cancel chip
- PR6d: wave-complete toast + 5-bucket summary + [Retry failed]
- PR6e: reflog drawer

## Open rubber-duck items still pending

Carry-forwards (V1.0 backlog, not gating PR6):
- DTO original_prompt_id nullability (V1.0 ships nullable per spec)
- PR1 backward-compat fallback corpus verification (needs prod DB)
- PR4e cancelQueued-race (documented; PR6 surface)
- dispatch.ts 50% branch coverage (pre-existing PR4c2)
- Spec drift docs (end-of-V1.0 pass): drain-outside-lock, retry-
  failed-in-shim, LockAcquireResult DU, SendCard inline lastError,
  per-child Pick toggle (Q2 = keep + amend), 💬 View raw response
  dropped (Q1), V1.1 stubs skipped (Q3)
- Stack-collapse persistence across un-stackable transitions (PR6)
- Auto-collapse seed one-shot per tree.id (PR6)
- TreeCanvas synchronous-setState-during-render (track)
- Playwright smoke test for edge-chip e2e flow (deferred)

End-of-V1.0 doc-pass list (5 spec amendments to land):
- §2.4 per-child Pick toggle (Q2)
- §2.2 drop 💬 View raw response (Q1)
- §2.4 no V1.1 stub reservation (Q3)
- §2.2 SendCard inline lastError surface
- §2.2 / runtime contract notes for V1.0 editing affordances
